# Make certain pages load faster