--- a/cypress/e2e/onboarding.cy.js
+++ b/cypress/e2e/onboarding.cy.js
@@ -242,7 +242,7 @@ describe("onboarding spec", () => {
         cy.contains("Foundation Partner").click({ force: true });
         cy.get("body").click(0, 0); // close dropdwon
 
-        cy.get('button[type="submit"]').click();
+        cy.get('button[type="submit"]').click({force: true});
         cy.url({ timeout: 60000 }).should(
           "include",
           "/welcome/existing-member/add-profile-info"