When pairing with @taylorz I noticed some pages were frustratingly slow. These are some time changes tested locally on my machine, pointing to production API.
 
`/ssj/[workflow]/startup`: 889 ms -> 259 ms (3.4x faster)
`/ssj/[workflow]/visioning`: 306 ms -> 119 ms (2.5x faster)
`/ssj/[workflow]/visioning`: 722 ms -> 202 ms (3.5x faster)
`/ssj/[workflow]/milestones`: 1.6 s -> 252 ms (6.3x faster)